+++ code/kaf_profiti/models/__init__.py
@@ -1,6 +1,7 @@
 """Model components for KAFNet-ProFITi."""
 
 from .gaussian_head import GaussianHead
+from .baselines import GRUDGaussian, GRUDGaussianConfig, TCNGaussian, TCNGaussianConfig
 from .kaf_gaussian import KAFGaussian, KAFGaussianConfig
 from .kaf_profiti import KAFProFITi, KAFProFITiConfig
 from .kafnet_encoder import KAFNetEncoder
@@ -9,6 +10,8 @@ from .query_condition_adapter import QueryConditionAdapter
 
 __all__ = [
     "GaussianHead",
+    "GRUDGaussian",
+    "GRUDGaussianConfig",
     "KAFGaussian",
     "KAFGaussianConfig",
     "KAFProFITi",
@@ -16,4 +19,6 @@ __all__ = [
     "KAFNetEncoder",
     "ProFITiFlowHead",
     "QueryConditionAdapter",
+    "TCNGaussian",
+    "TCNGaussianConfig",
 ]